Modified SearchFacet component to not use shadow DOM. Created new es-search.css for search interface styles.
