+++ _includes/head/head.html
@@ -16,6 +16,7 @@
 <script defer src="{{ '/assets/lib/fontawesome-free/js/all.min.js' | absolute_url }}"></script>
 {% if layout.gallery == true %}<link rel="stylesheet" href="{{ '/assets/lib/lightgallery/css/lightgallery.min.css' | relative_url }}" type="text/css">{%- endif -%}
 {% if page.layout == 'data' %}<link rel="stylesheet" type="text/css" href="{{ '/assets/lib/datatables/datatables.min.css' | absolute_url }}">{%- endif -%}
+{% if page.layout == 'search' or page.layout == 'multi-collection-search' %}<link rel="stylesheet" type="text/css" href="{{ '/assets/css/es-search.css' | absolute_url }}">{%- endif -%}
 {% if page.layout == 'map' %}<link rel="stylesheet" href="{{ '/assets/lib/leaflet/leaflet.css' | absolute_url }}">
 {% if site.data.theme.map-search == true %}<link rel="stylesheet" href="{{ '/assets/lib/leaflet/leaflet.fusesearch.css' | absolute_url }}">{% endif %}
 {% if site.data.theme.map-cluster == true %}<link rel="stylesheet" href="{{ '/assets/lib/leaflet/MarkerCluster.css' | absolute_url }}">
